market-spine-full/pyproject.toml
[project]
name = "market-spine-full"
version = "1.0.0"
description = "Production-grade Market Spine analytics pipeline with Kubernetes support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "psycopg[binary,pool]>=3.2.0",
    # Task queue
    "celery[redis]>=5.4.0",
    # HTTP client
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    # Storage
    "boto3>=1.35.0",
    # Configuration
    "pydantic-settings>=2.6.0",
    # CLI
    "typer>=0.15.0",
    # Observability
    "prometheus-client>=0.21.0",
    "structlog>=24.4.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-celery>=0.49b0",
    "opentelemetry-instrumentation-psycopg>=0.49b0",
    "opentelemetry-exporter-otlp>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
spine = "market_spine.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/market_spine"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning:psycopg_pool.*",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "C4", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
